backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pymongo>=4.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
google-search-results>=2.4.2
pymilvus>=2.4.3
sentence-transformers>=3.0.1
torch>=2.3.1
pypdf2>=3.0.1
python-multipart>=0.0.9
supabase>=2.5.0
marshmallow<4
openai>=1.30.0
weasyprint>=60.0
redis>=5.0.0
pyjwt[crypto]>=2.8.0
websockets>=12.0
groq>=0.9.0
deepgram-sdk>=3.2.0
